api/requirements.txt
azure-functions
openai==0.28.1
python-dotenv

# Testing Dependencies
pytest
pytest-mock
